Save Game Array Start
Added found/unfound string array to game instance and saves, and integrated it with mainsession UI menu.  Needs a link adding after scanning using AR tool to call the UpdateArray Function under the game instance, to change found/unfound status of the artwork scanned.  Takes row number of the artwork from the data table as the index.  Also needs save call adding after updating the list to save that change to the savefile as well as to the current game instance